requirements.txt

flask
python-telegram-bot
requests
aiohttp
python-dotenv
telegram
psutil
asyncio
